iextrading4j-client/src/test/java/pl/zankowski/iextrading4j/client/sse/request/forex/CurrencyRatesSseRequestBuilderTest.java
package pl.zankowski.iextrading4j.client.sse.request.forex;

import org.junit.Test;
import pl.zankowski.iextrading4j.api.forex.CurrencyRate;
import pl.zankowski.iextrading4j.client.sse.manager.SseRequest;

import javax.ws.rs.core.GenericType;
import java.util.List;

import static org.assertj.core.api.Assertions.assertThat;
import static org.assertj.core.api.Assertions.entry;

public class CurrencyRatesSseRequestBuilderTest {

    @Test
    public void shouldSuccessfullyCreateRequest() {
        final String symbol = "USDJPY";

        final SseRequest<List<CurrencyRate>> request = new CurrencyRatesSseRequestBuilder()
                .withSymbol(symbol)
                .build();

        assertThat(request.getPath()).isEqualTo("/forex");
        assertThat(request.getResponseType()).isEqualTo(new GenericType<List<CurrencyRate>>() {});
        assertThat(request.getPathParams()).isEmpty();
        assertThat(request.getQueryParams()).contains(entry("nosnapshot", "false"), entry("symbols", symbol));
    }

}
